Refaktoriser Observations-klasser til eget modul
ObservationsType, Observation og nedarvende klasser flyttet til
fire.api.model.observationer for at øge overskueligheden af koden.
