+++ fire/api/model/__init__.py
@@ -130,3 +130,4 @@ class RegisteringTidObjekt(DeclarativeBase):
 from .geometry import *
 from .punkttyper import *
 from .sagstyper import *
+from .observationer import *
